Restricted driving working but no comments

+++ Algorithms.h
@@ -12,7 +12,6 @@
  * @brief Prints the path in the correct order
  *
  * @param v The vector of ints representing the id of each location in the path.
- * The path is in reversed order (from destination to origin)
  */
 void printPath(const std::vector<int> &v);
 
@@ -23,12 +22,12 @@ void printPath(const std::vector<int> &v);
  * @param g A pointer to the graph that has the origin and destination Vertex
  * @param origin The id of the origin vertex of the path that is being calculated
  * @param dest The id of the destination vertex of the path that is being calculated
- * @param time A reference to an int where the total time of the path is being held
+ * @param time A reference to a double where the total time of the path is being held
  *
- * @return A vector of ints representing the id of each location in the path. The
- * path is in reversed order (from destination to origin)
+ * @return A vector of ints representing the id of each location in the path.
  */
-std::vector<int> getPath(Graph * g, const int &origin, const int &dest, int &time);
+
+std::vector<int> getPath(Graph * g, const int &origin, const int &dest, double &time);
